pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelbeans"
version = "0.1.0"
description = "Photo to perler bead pattern converter with ΔE2000 color quantization"
requires-python = ">=3.10"
authors = [{ name = "PixelBeans" }]
dependencies = [
    "numpy>=1.24",
    "Pillow>=10.0",
    "opencv-python-headless>=4.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "colour-science>=0.4",
]

[project.scripts]
pixelbeans = "pixelbeans.cli:main"

[tool.setuptools.packages.find]
include = ["pixelbeans*"]

[tool.setuptools.package-data]
pixelbeans = ["../palettes/*.json"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
